requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.10.8
Django==4.0.1
django-bootstrap5==21.3
django-ckeditor==6.2.0
django-crispy-forms==1.14.0
django-elasticsearch-dsl==7.2.2
django-js-asset==1.2.2
django-taggit==2.0.0
elasticsearch==7.16.3
elasticsearch-dsl==7.4.0
Markdown==3.3.6
Pillow==9.0.0
psycopg2-binary==2.9.3
python-dateutil==2.8.2
python-decouple==3.5
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
urllib3==1.26.8
